Add new score and config variable ::ProblemResolver::CancelRemovalScore
Add new score for partial solutions cancelling the removal of a package and
corresponding config variable ::ProblemResolver::CancelRemovalScore

Resolver actions cancelling the removal (or purge) of a package requested to
be removed/purged were not being penalised, now they are (default -300).

This was missing compared to penalisations for other actions like
RemoveScore, when removing packages not marked for removal.  Now, solutions
of the resolver cancelling explicitly requested removals are also penalised.