Actions involving unfulfilled Recommends invoke the resolver (Closes: #819636)
With APT::Install-Recommends="true" (the default), actions resulting in
unavailable Recommends (e.g., the recommended package is upgraded instead of
remaining in the required version) invoke the resolver, for the user to take
decision.

This is because now the resolver monitors situation when the
PolicyBrokenCount is not zero, before it was only taken BrokenCount into
account (it was the only count available when the code was set up).