Enhance infer_reason() to explain more cases of pkg_unused_remove (Closes: #266061)
Enhance infer_reason() to explain more cases where package is removed
because of automatically installed and unused (one case related to #819636).

The case related with #819636 is when the version of the depended package
(the one for which we look the reason of removal) changes because e.g. it is
upgraded, so maybe it doesn't fulfill the original dependency (for example
because the package depending on it it depends on an exact or less-than
version, as in the bug report).

The second case (#266061) was missing the explanation because it didn't take
into account reverse-depends of virtual packages provided by the package
being inspected.