"apply solution" approves broken Recommends in current solution (part of #819636)
This is part of the work to fix #819636.

The user was able to acknowledge the breakage because the resolver gets invoked
whenever recommendations are broken (if APT::Install-Recommends was true, and it
is by default).  That wasn't happening before changes to solve #819636, when
breakages to recommends were silently accepted.

But when invoking the resolver also because of "policy broken" packages, and
since the resolver gets discarded and created again as part of applying the
solution (package state changes in aptcache), we need a way to break out of the
loop.

With this change, the broken soft dependencies (like Recommends) from the
currently selected solution are approved when "apply solution" is taken, and
reinstated after the resolver is created again.

A problem might arise when the user changes its mind and rejects previoulsy
approved recommends, but this problem should be much harder to hit than the one
before, which was already rare.  It would be solved also by not having "policy
broken" into account in subsequent calls (there is some support to do this now),
but this would make the resolver to ignore subsequent broken Recommends between
the first "apply solution" and

In the end, only so much can be done without completely forbidding installing
packages without its recommends.  This is necessary sometimes when the versions
in the Debian archive are out of sync or temporarily broken, or if the user
wants to deselect a recommend (before installing) which is being pulled as part
of installing a new package, without having to disable APT::Install-Recommends.

I suspect that many users rely on the current behaviour present for years, so
being completely inflexible with missing recommends is probably a worse solution
overall than the current imperfect implementation.