Distinguish Debian-specific and upstream upgrades
When upgrading packages, aptitude users might like to know whether each
upgrade represents a newer upstream version of the package, or is merely
a Debian-specific change.  This can be determined by comparing the
current and candidate versions in the rightmost two columns of the
package listing, but I thought it might be easier to see at a glance if
these two kinds of upgrades were colored differently in the UI.

This commit makes upgrades that represent a newer upstream version
appear in bold, while upgrades that represent a Debian revision bump
appear as before.