[curses] Do not fail immediately when some package lists cannot be updated (Closes: #834790)
[curses] Do not fail immediately and attempt to continue when some package lists
cannot be updated (Closes: #834790)

The bug, in which the update of package lists stopped and the interface was
left at the "Loading cache" stage indefinitely, seems to be due to changes
in behaviour of apt or some strange interaction, because the relevant
sections of aptitude have not changed in a long time.

This is fixed now by not being strict with this kind of errors (to cope with
what seems to be a new behaviour from libapt -- to not report that the
update can proceed)... Hopefully this change will not create other problems.