Use a more strict mask for comparison of dependency operators (Closes: #836567, #849370)
Use a more strict mask for comparison of dependency operators, thanks to
Aaron M. Ucko for the initial patch (Closes: #836567, #849370)

Recently apt started to use more values in an enum used for dependency
operator comparisons, which causes some problems because aptitude's code
from many years ago, which did not expect new values.

Using a stricter mask to capture only the operators, as per the
documentation of the enum (lower 4 bits).